working brainfuck interpreter written in QUARTZ./quartz ./brainfuck.qz  all known errors are fixed. all known mem leaks and problems are fixed. detected a problem when you want to do (cast<Command>(self.commands.get(i))()

+++ qcc/token.h
@@ -73,7 +73,7 @@ typedef enum {
 typedef struct {
     TokenKind kind;
     const char* start;
-    uint8_t length;
+    int length;
     uint32_t line;
     uint32_t column;
 } Token;
